backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
midtransclient
python-multipart
Pillow
